% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dplyr_pithy.R
\name{dplyr_pithy}
\alias{dplyr_pithy}
\alias{f_pithy}
\alias{f_pithy_}
\alias{filter_pithy}
\alias{filter_pithy_}
\alias{m_pithy}
\alias{m_pithy_}
\alias{mutate_pithy}
\alias{mutate_pithy_}
\alias{s_pithy}
\alias{s_pithy_}
\alias{select_pithy}
\alias{select_pithy_}
\alias{t_pithy}
\alias{t_pithy_}
\alias{transmute_pithy}
\alias{transmute_pithy_}
\title{Filter, Select, Mutate, or Transmute Data Prior to Plotting}
\usage{
filter_pithy(.data, ..., .pithargs = list())

f_pithy(.data, ..., .pithargs = list())

filter_pithy_(.data, ..., .pithargs = list())

f_pithy_(.data, ..., .pithargs = list())

mutate_pithy(.data, ..., .pithargs = list())

m_pithy(.data, ..., .pithargs = list())

mutate_pithy_(.data, ..., .pithargs = list())

m_pithy_(.data, ..., .pithargs = list())

select_pithy(.data, ..., .pithargs = list())

s_pithy(.data, ..., .pithargs = list())

select_pithy_(.data, ..., .pithargs = list())

s_pithy_(.data, ..., .pithargs = list())

transmute_pithy(.data, ..., .pithargs = list())

t_pithy(.data, ..., .pithargs = list())

transmute_pithy_(.data, ..., .pithargs = list())

t_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to
\code{\link[dplyr]{filter}},
\code{\link[dplyr]{filter_}},
\code{\link[dplyr]{select}},
\code{\link[dplyr]{select_}},
\code{\link[dplyr]{mutate}},
\code{\link[dplyr]{mutate_}},
\code{\link[dplyr]{transmute}}, or
\code{\link[dplyr]{transmute_}},}

\item{.pithargs}{List of arguments to pass to \code{\link{pithy}}.}
}
\description{
These functions use functions from \code{\link[dplyr]{dplyr}} produce
plot from a modified data set before returning the unmodified data
set.
}
\details{
These functions let data.frame, tbl, or tbl_df objects be modified
prior to producing plots with \code{\link{pithy}}.  Just like pithy,
these functions return the unmodified data set.  \code{filter_pithy}
calls \code{\link[dplyr]{filter}} prior to calling \code{pithy}.
\code{select_pithy} calls \code{\link[dplyr]{select}} prior to calling
\code{pithy}.  \code{mutate_pithy} calls \code{\link[dplyr]{mutate}}
prior to calling \code{pithy}.  \code{transmute_pithy} calls
\code{\link[dplyr]{transmute}} prior to calling pithy.

\code{f_pithy}, \code{s_pithy}, \code{m_pithy}, and \code{t_pithy}
are shorter synonyms for \code{filter_pithy}, \code{select_pithy},
\code{mutate_pithy}, and \code{transmute_pithy}, respectively.

\code{filter_pithy_}, \code{f_pithy_}, \code{select_pithy_},
\code{s_pithy_}, \code{mutate_pithy_}, \code{m_pithy_},
\code{transmute_pithy_}, and \code{t_pithy_} are similar to the
corresponding functions without the trailing underscore, but call
\code{\link[dplyr]{filter_}}, \code{\link[dplyr]{select_}},
\code{\link[dplyr]{mutate_}}, and \code{\link[dplyr]{transmute_}}
instead of \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}},
\code{\link[dplyr]{mutate}}, and \code{\link[dplyr]{transmute}}.
}
\examples{
# filter_pithy -----
head(filter_pithy(cars, speed > 10))

# select_pithy -----
head(select_pithy(cars, dist))

# mutate_pithy
head(mutate_pithy(cars, ratio = dist / speed))

# transmute_pithy
head(cars, ratio = dist / speed))
}

