% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_pithy.R
\name{filter_pithy}
\alias{f_pithy}
\alias{filter_pithy}
\title{Filter rows, generate a \code{pith}, then return the original data set}
\usage{
filter_pithy(.data, ..., .pithargs = list())

f_pithy(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{filter}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Filter rows, generate a \code{pith}, then return the original data set
}
\examples{
head(filter_pithy(iris, Species == "setosa"))
head(f_pithy(iris[, c(3,5)], Petal.Length > 2))
}
\seealso{
\code{\link{pithy}}, \code{\link{select_pithy}}, \code{\link{mutate_pithy}}, \code{\link{transmute_pithy}}
}

