% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_pithy_.R
\name{filter_pithy_}
\alias{f_pithy_}
\alias{filter_pithy_}
\title{Filter rows, generate a \code{pith}, then return the original data set}
\usage{
filter_pithy_(.data, ..., .pithargs = list())

f_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{filter_}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Filter rows, generate a \code{pith}, then return the original data set
}
\section{Functions}{
\itemize{
\item \code{filter_pithy_}: Just like f_pithy_, but with more typing.

\item \code{f_pithy_}: Just like filter_pithy_, but with less typing.
}}
\examples{
head(filter_pithy_(iris, "Species == 'setosa'"))
head(f_pithy_(iris[, c(3,5)], "Petal.Length > 2"))
}
\seealso{
\code{\link{pithy}}, \code{\link{select_pithy_}}, \code{\link{mutate_pithy_}}, \code{\link{transmute_pithy_}}
}

