% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutate_pithy.R
\name{mutate_pithy}
\alias{mutate_pithy}
\alias{mutate_pithy_}
\title{Mutate Data Prior to Plotting}
\usage{
mutate_pithy(.data, ..., .pithargs = list())

mutate_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to
\code{\link[dplyr]{mutate}} or
\code{\link[dplyr]{mutate_}},}

\item{.pithargs}{List of arguments to pass to \code{\link{pithy}}.}
}
\value{
\code{.data}, unmodified.
}
\description{
These functions call \code{\link[dplyr]{mutate}} or
\code{\link[dplyr]{mutate_}} before calling \code{\link{pithy}}.
}
\details{
These functions let data.frame, tbl, or tbl_df objects be modified
prior to producing plots with \code{\link{pithy}}.  Just like pithy,
these functions return the unmodified data set.  \code{mutate_pithy}
calls \code{\link[dplyr]{mutate}} prior to calling \code{\link{pithy}}.
\code{mutate_pithy_} calls \code{\link[dplyr]{mutate_}} prior to
calling \code{\link{pithy}}.
}
\examples{
# mutate_pithy
head(mutate_pithy(cars, ratio = dist / speed))
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link{pithy}},
\code{\link{filter_pithy}}, \code{\link{select_pithy}},
\code{\link{transmute_pithy}}
}

