% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutate_pithy_.R
\name{mutate_pithy_}
\alias{m_pithy_}
\alias{mutate_pithy_}
\title{Mutate variables, generate a \code{pith}, then return the original data set}
\usage{
mutate_pithy_(.data, ..., .pithargs = list())

m_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{mutate_}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Mutate variables, generate a \code{pith}, then return the original data set
}
\section{Functions}{
\itemize{
\item \code{mutate_pithy_}: Just like m_pithy_, but with more typing.

\item \code{m_pithy_}: Just like mutate_pithy_, but with less typing.
}}
\examples{
head(mutate_pithy_(iris["Sepal.Length"], "SL2 = Sepal.Length^2"))
head(m_pithy_(iris["Petal.Width"], "Wide_Petal = Petal.Width > mean(Petal.Width)"))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy_}}, \code{\link{select_pithy_}}, \code{\link{transmute_pithy_}}
}

