% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.array.R
\name{pith.array}
\alias{pith.array}
\title{Generate a \code{pith} class object from a \code{\link[base]{array}}}
\usage{
\method{pith}{array}(x, freq = TRUE, plot = TRUE, xname = NULL,
  histargs = list(), ...)
}
\arguments{
\item{x}{A array.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{histargs}{List of additional arguments ot pass to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} object, invisibly.
}
\description{
Arrays are collapsed into vectors.
}
\examples{
X <- array(sample(c("A", "B", "B", "C", NA),
  size = 100, replace = TRUE),
  c(5, 4, 5))
pith(X)
XX <- array(rpois(200, 500), c(25, 4, 2))
pith(XX)
}

