% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.character.R
\name{pith.character}
\alias{pith.character}
\title{Generate a \code{pith} class object from a \code{\link[base]{character}}}
\usage{
\method{pith}{character}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{A character vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} class object, invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{character}}
}
\examples{
X <- sample(
  c(LETTERS[c(1, 2, 2, 3, 3, 3)], NA),
  size = 50, replace = TRUE)
pith(X)
}

