% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.data.frame.R
\name{pith.data.frame}
\alias{pith.data.frame}
\title{Generate a \code{pith} class object from a \code{\link[base]{data.frame}}}
\usage{
\method{pith}{data.frame}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{A data.frame.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pithlist} class object, invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{data.frame}}
}
\examples{
pith(iris)
}

