% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.integer.R
\name{pith.integer}
\alias{pith.integer}
\title{Generate a \code{pith} class object from a \code{\link[base]{integer}}}
\usage{
\method{pith}{integer}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{An integer vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} class object, invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{integer}}
}
\details{
If the range of integers in \code{x} is greater than 25, then \code{x} is treated as a numeric
vector, otherwise \code{x} is treated as a factor vector.
}
\examples{
pith(rpois(50, 5))
pith(rpois(50, 500))
}

