% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.list.R
\name{pith.list}
\alias{pith.list}
\title{Generate a \code{pith} class object from a \code{\link[base]{list}}}
\usage{
\method{pith}{list}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{A list.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pithlist} object (a list of \code{pith} objects), invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{list}}
}
\examples{
pith(list(A = rnorm(50), B = rpois(50, 5)))
}

