% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.logical.R
\name{pith.logical}
\alias{pith.logical}
\title{Generate a \code{pith} class object from a \code{\link[base]{logical}}}
\usage{
\method{pith}{logical}(x, freq = TRUE, plot = TRUE, xname = NULL,
  histargs = list(), ...)
}
\arguments{
\item{x}{A logical vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{histargs}{List. Ignored.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{logical}}
}

