% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.matrix.R
\name{pith.matrix}
\alias{pith.matrix}
\title{Generate a \code{pith} class object from a \code{\link[base]{matrix}}}
\usage{
\method{pith}{matrix}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{A matrix.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} object, invisibly.
}
\description{
It is difficult to infer, from a matrix object alone, whether it would
be better to split the matrix into rows or columns, so the matrix isn't
split up at all.  Instead it is converted to a vector.
}
\examples{
X <- matrix(sample(c("A", "B", "B", "C", NA),
  size = 100, replace = TRUE),
  ncol = 10)
pith(X)
XX <- matrix(rpois(200, 500), ncol = 10)
pith(XX)
}

