% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.numeric.R
\name{pith.numeric}
\alias{pith.numeric}
\title{Generate a \code{pith} class object from a \code{\link[base]{numeric}}}
\usage{
\method{pith}{numeric}(x, freq = TRUE, plot = TRUE, xname = NULL,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{A factor vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{breaks}{Passed to \code{\link[graphics]{hist}}.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} class object, invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{numeric}}
}
\examples{
X <- rnorm(80)
pith(X)
XX <- c(X, rep(NA, 20))
pith(XX)
}

