\name{pithr-package}
\alias{pithr-package}
\alias{pithr}
\docType{package}
\title{
Produce quick graphical data summaries!
}
\description{
Piths are simple summaries of data which include factor frequencies or a histogram as well as a frequency of NA, if applicable.
}
\details{
\tabular{ll}{
Package: \tab pithr\cr
Type: \tab Package\cr
Version: \tab 0.2.4\cr
Date: \tab 2015-03-16\cr
License: \tab GPL>=2\cr
}
}
\author{
Nick Salkowski

Maintainer: Nick Salkowski <nicksalkowskiphd@gmail.com>
}
\keyword{ package }
