% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pithy.R
\name{pithy}
\alias{pithy}
\title{Generate a \code{\link{pith}}, but return the original \code{x}}
\usage{
pithy(x, xname = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{xname}{Character string describing the factor vector.}

\item{...}{Additional arguments passed to \code{\link{pith}}}
}
\value{
\code{x}, unaltered.
}
\description{
Generate a \code{\link{pith}}, but return the original \code{x}
}
\examples{
X <- sample(
  c(LETTERS[c(1, 2, 2, 3, 3, 3)], NA),
  size = 50, replace = TRUE)
pithy(X)
head(pithy(iris))
}

