% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.pithlist.R
\name{plot.pithlist}
\alias{plot.pithlist}
\title{Plot a \code{pithlist}}
\usage{
\method{plot}{pithlist}(x, ...)
}
\arguments{
\item{x}{A \code{pithlist}}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
\code{NULL}, invisibly.
}
\description{
Plot a \code{pithlist}
}
\seealso{
\code{\link{plot.pith}}
}

