% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pselect.R
\name{pselect}
\alias{pselect}
\title{Select and Pith}
\usage{
pselect(.data, ..., .pith = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{A comma separated list of unquoted expressions passed to \code{\link[dplyr]{select}}.}

\item{.pith}{A list of arguments to pass to \code{\link{pith}}}
}
\value{
A data.frame, tbl, or tbl_df.
}
\description{
\code{pselect} first runs \code{\link[dplyr]{select}} on a data set, runs \code{\link{pith}}, then returns the selected columns from the data set.
}

