% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_pithy.R
\name{select_pithy}
\alias{s_pithy}
\alias{select_pithy}
\title{Select variables, generate a \code{pith}, then return the original data set}
\usage{
select_pithy(.data, ..., .pithargs = list())

s_pithy(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{select}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Select variables, generate a \code{pith}, then return the original data set
}
\examples{
head(select_pithy(iris, Sepal.Length))
head(s_pithy(iris, Petal.Length, Petal.Width))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy}}, \code{\link{mutate_pithy}}, \code{\link{transmute_pithy}}
}

