% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_pithy_.R
\name{select_pithy_}
\alias{s_pithy_}
\alias{select_pithy_}
\title{Select variables, generate a \code{pith}, then return the original data set}
\usage{
select_pithy_(.data, ..., .pithargs = list())

s_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{select_}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Select variables, generate a \code{pith}, then return the original data set
}
\section{Functions}{
\itemize{
\item \code{select_pithy_}: Just like s_pithy_, but with more typing.

\item \code{s_pithy_}: Just like select_pithy_, but with less typing.
}}
\examples{
head(select_pithy_(iris, "Sepal.Length"))
head(s_pithy_(iris, "Petal.Length", "Petal.Width"))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy_}}, \code{\link{mutate_pithy_}}, \code{\link{transmute_pithy_}}
}

