% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transmute_pithy.R
\name{transmute_pithy}
\alias{t_pithy}
\alias{transmute_pithy}
\title{Transmute variables, generate a \code{pith}, then return the original data set}
\usage{
transmute_pithy(.data, ..., .pithargs = list())

t_pithy(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{transmute}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Transmute variables, generate a \code{pith}, then return the original data set
}
\section{Functions}{
\itemize{
\item \code{transmute_pithy}: Just like t_pithy, but with more typing.

\item \code{t_pithy}: Just like transmute_pithy, but with less typing.
}}
\examples{
head(transmute_pithy(iris, SL2 = Sepal.Length^2))
head(t_pithy(iris, Wide_Petal = Petal.Width > mean(Petal.Width)))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy}}, \code{\link{select_pithy}}, \code{\link{mutate_pithy}}
}

