% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transmute_pithy_.R
\name{transmute_pithy_}
\alias{t_pithy_}
\alias{transmute_pithy_}
\title{Transmute variables, generate a \code{pith}, then return the original data set}
\usage{
transmute_pithy_(.data, ..., .pithargs = list())

t_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{transmute_}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Transmute variables, generate a \code{pith}, then return the original data set
}
\examples{
head(transmute_pithy_(iris, "SL2 = Sepal.Length^2"))
head(t_pithy_(iris, "Wide_Petal = Petal.Width > mean(Petal.Width)"))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy_}}, \code{\link{select_pithy_}}, \code{\link{mutate_pithy_}}
}

