% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_pithy.R
\name{filter_pithy}
\alias{filter_pithy}
\alias{filter_pithy_}
\title{Filter Data Prior to Plotting}
\usage{
filter_pithy(.data, ..., .pithargs = list())

filter_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to
\code{\link[dplyr]{filter}} or
\code{\link[dplyr]{filter_}},}

\item{.pithargs}{List of arguments to pass to \code{\link{pithy}}.}
}
\value{
\code{.data}, unmodified.
}
\description{
These functions call \code{\link[dplyr]{filter}} or
\code{\link[dplyr]{filter_}} before calling \code{\link{pithy}}.
}
\details{
These functions let data.frame, tbl, or tbl_df objects be modified
prior to producing plots with \code{\link{pithy}}.  Just like pithy,
these functions return the unmodified data set.  \code{filter_pithy}
calls \code{\link[dplyr]{filter}} prior to calling \code{\link{pithy}}.
\code{filter_pithy_} calls \code{\link[dplyr]{filter_}} prior to
calling \code{\link{pithy}}.
}
\examples{
# filter_pithy -----
head(filter_pithy(cars, speed > 10))
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link{pithy}},
\code{\link{select_pithy}}, \code{\link{mutate_pithy}},
\code{\link{transmute_pithy}}
}

