% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutate_pithy.R
\name{mutate_pithy}
\alias{m_pithy}
\alias{mutate_pithy}
\title{Mutate variables, generate a \code{pith}, then return the original data set}
\usage{
mutate_pithy(.data, ..., .pithargs = list())

m_pithy(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{mutate}}}

\item{.pithargs}{List of arguments to pass to pithy.}
}
\value{
Unaltered \code{\.data}
}
\description{
Mutate variables, generate a \code{pith}, then return the original data set
}
\section{Functions}{
\itemize{
\item \code{mutate_pithy}: Just like m_pithy, but with more typing.

\item \code{m_pithy}: Just like mutate_pithy, but with less typing.
}}
\examples{
head(mutate_pithy(iris["Sepal.Length"], SL2 = Sepal.Length^2))
head(m_pithy(iris["Petal.Width"], Wide_Petal = Petal.Width > mean(Petal.Width)))
}
\seealso{
\code{\link{pithy}}, \code{\link{filter_pithy}}, \code{\link{select_pithy}}, \code{\link{transmute_pithy}}
}

