% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.character.R
\name{pith.character}
\alias{pith.character}
\title{Generate a \code{pith} class object from a \code{\link[base]{character}}}
\usage{
\method{pith}{factor}(x, freq = TRUE, plot = TRUE, xname = NULL,
  histargs = list(), ...)
}
\arguments{
\item{x}{A character vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{histargs}{List. Ignored.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{character}}
}

