% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.factor.R
\name{pith.factor}
\alias{pith.factor}
\title{Generate a \code{pith} class object from a \code{\link[base]{factor}}}
\usage{
\method{pith}{factor}(x, freq = TRUE, plot = TRUE, xname = NULL,
  histargs = list(), ...)
}
\arguments{
\item{x}{A factor vector.}

\item{freq}{Logical. If \code{TRUE}, frequencies are plotted instead of proportions.}

\item{plot}{Logical. If TRUE, the \code{pith} is plotted.}

\item{xname}{Character string describing the factor vector.}

\item{histargs}{List. Ignored.}

\item{...}{Additional arguments passed to \code{\link{plot.pith}}}
}
\value{
A \code{pith} class object, invisibly.
}
\description{
Generate a \code{pith} class object from a \code{\link[base]{factor}}
}
\examples{
X <- factor(sample(
  c(LETTERS[c(1, 2, 2, 3, 3, 3)], NA),
  size = 50, replace = TRUE))
pith(X)
}

