% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.R
\name{pith}
\alias{pith}
\title{Generate a \code{pith} class object}
\usage{
pith(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\value{
A \code{pith} or \code{pithlist} class object.
}
\description{
Generate a \code{pith} class object
}
\seealso{
\code{\link{pith.factor}}, \code{\link{pith.character}}, \code{\link{pith.logical}}, \code{\link{pith.integer}}, \code{\link{pith.numeric}}, \code{\link{pith.list}}, \code{\link{pith.data.frame}}
}

