% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pith.R
\name{pith}
\alias{pith}
\title{Generate a \code{pith} class object}
\usage{
pith(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\description{
Generate a \code{pith} class object
}

