% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pithy.R
\name{pithy}
\alias{pithy}
\title{Generate a Pith, but return the original \code{x}}
\usage{
pithy(x, xname = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{xname}{Character string describing the factor vector.}

\item{...}{Additional arguments passed to \code{\link{pith}}}
}
\value{
\code{x}, unaltered.
}
\description{
\code{pithy} calls \code{\link{pith}}, but returns
the original data, rather than a \code{pith} class object.
}
\examples{
X <- sample(
  c(LETTERS[c(1, 2, 2, 3, 3, 3)], NA),
  size = 50, replace = TRUE)
pithy(X)
head(pithy(iris))
}
\seealso{
\code{\link{pith}}, \code{\link{filter_pithy}},
\code{\link{select_pithy}}, \code{\link{mutate_pithy}},
\code{\link{transmute_pithy}}
}

