% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.pith.R
\name{plot.pith}
\alias{plot.pith}
\title{Plot a pith class object}
\usage{
\method{plot}{pith}(x, freq = TRUE, main = NULL, xlab = NULL,
  ylab = NULL, col = c("#619CFF", "#F8766D"), border = NA,
  shrink = TRUE, ...)
}
\arguments{
\item{x}{A pith.}

\item{freq}{Logical.  If \code{TRUE}, frequencies will be plotted instead of
proportions or densities.}

\item{main}{Character string. Main title, default is \code{x$xname}}

\item{xlab}{Character string. X-axis label.}

\item{ylab}{Character string. Y-axis label.}

\item{col}{Character vector.  Up to two fill colors may be specified, the
first for the nonmissing data bars and the second for the NA bar.  If only
one color is specified it will be used for all bars.}

\item{border}{Character vector.  Up to two border colors may be specified,
the first for the nonmissing data bars and the second for the NA bar.  If
only one color is specified it will be used for all bars.}

\item{shrink}{Logical.  If TRUE, factor labels will shrink to fit the width
of the bar.  Ignored
if \code{las} is either 2 or 3.}

\item{...}{Additional plot parameters.  See \code{\link[graphics]{par}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Plot a pith class object
}

