% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_pithy.R
\name{select_pithy}
\alias{select_pithy}
\alias{select_pithy_}
\title{Select Data Prior to Plotting}
\usage{
select_pithy(.data, ..., .pithargs = list())

select_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to
\code{\link[dplyr]{select}} or
\code{\link[dplyr]{select_}},}

\item{.pithargs}{List of arguments to pass to \code{\link{pithy}}.}
}
\value{
\code{.data}, unmodified.
}
\description{
These functions call \code{\link[dplyr]{select}} or
\code{\link[dplyr]{select_}} before calling \code{\link{pithy}}.
}
\details{
These functions let data.frame, tbl, or tbl_df objects be modified
prior to producing plots with \code{\link{pithy}}.  Just like pithy,
these functions return the unmodified data set.  \code{select_pithy}
calls \code{\link[dplyr]{select}} prior to calling \code{\link{pithy}}.
\code{select_pithy_} calls \code{\link[dplyr]{select_}} prior to
calling \code{\link{pithy}}.
}
\examples{
# select_pithy -----
head(select_pithy(cars, dist))
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link{pithy}},
\code{\link{filter_pithy}}, \code{\link{mutate_pithy}},
\code{\link{transmute_pithy}}
}

