% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transmute_pithy.R
\name{transmute_pithy}
\alias{transmute_pithy}
\alias{transmute_pithy_}
\title{Transmute Data Prior to Plotting}
\usage{
transmute_pithy(.data, ..., .pithargs = list())

transmute_pithy_(.data, ..., .pithargs = list())
}
\arguments{
\item{.data}{A data.frame, tbl, or tbl_df}

\item{...}{Additional arguments to pass to
\code{\link[dplyr]{transmute}} or
\code{\link[dplyr]{transmute_}},}

\item{.pithargs}{List of arguments to pass to \code{\link{pithy}}.}
}
\value{
\code{.data}, unmodified.
}
\description{
These functions call \code{\link[dplyr]{transmute}} or
\code{\link[dplyr]{transmute_}} before calling \code{\link{pithy}}.
}
\details{
These functions let data.frame, tbl, or tbl_df objects be modified
prior to producing plots with \code{\link{pithy}}.  Just like pithy,
these functions return the unmodified data set.  \code{transmute_pithy}
calls \code{\link[dplyr]{transmute}} prior to calling \code{\link{pithy}}.
\code{transmute_pithy_} calls \code{\link[dplyr]{transmute_}} prior to
calling \code{\link{pithy}}.
}
\examples{
# transmute_pithy
head(transmute_pithy(cars, ratio = dist / speed))
}
\seealso{
\code{\link[dplyr]{transmute}}, \code{\link{pithy}},
\code{\link{filter_pithy}}, \code{\link{select_pithy}},
\code{\link{mutate_pithy}}
}

